corewars_core
Module load_file

Structs

DEFAULT_CONSTANTS

Field

Instruction

Metadata

Metadata about a Redcode program that is stored in the comments.

Offset

An absolute (non-negative) offset from the beginning of a core.

Program

A parsed Redcode program, which can be loaded into a core for execution

Warrior

The main public struct used to represent a Redcode warrior

Enums

AddressMode

Modifier

Opcode

PseudoOpcode

Value

Type Aliases

Instructions

LabelMap